selene: perf: Add in MTK APP TOUCH hints to INTERACTION hint
MTK APP TOUCH hints are triggered when an app is being used to
boost the responsiveness of the app.
But MTK APP TOUCH hints are non-existent on AOSP as they rely on MTK
boost framework. On the other hand AOSP provides handy boosting hints
by default with the AIDL interfaces.
These are available in the stock powerhal but are not enabled by
default because of the different approach of the MTK Boosting
framework.
Add these app touch hints to INTERACTION, making the UX even better.

Change-Id: I001249c08bf88af739ea8b86c58dcb05468cd6df
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Reviewed-on: https://review.statixos.com/c/android_device_xiaomi_cannon/+/7899
Signed-off-by: atharvnegi <atharvnegi2951@gmail.com>